Functional updates to 1.3 from master branch

+++ tests/CI/ci-reset-virtual-machine.sh
@@ -95,12 +95,20 @@ usage() {
 	echo "    --attempt-download [auto|yes|no] Should an OS image download be attempted at all?"
 	echo "                         (default: auto; default if only the option is specified: yes)"
 	echo "    --no-download        Alias to '--attempt-download no' to redeploy an existing image"
+	echo "    --no-delete          Do not clean up (just reboot VM if --no-download, or upgrade"
+	echo "                         in place if --with-overlayfs); a --with-delete reverses this"
+	echo "                         setting for a run if default is provided in configuration file"
 	echo "    --stop-only          end the script after stopping the VM and cleaning up"
 	echo "    --destroy-only       end the script after stopping the VM (container 'destroy')"
 	echo "    --deploy-only        end the script just before it would start the VM (skips apt-get too)"
 	echo "    --copy-host-users 'a b c'    Copies specified user or group account definitions"
 	echo "    --copy-host-groups 'a b c'   (e.g. for bind-mounted homes from host into the VM)"
 	echo "    --no-config-file     Forbid use in this run of a per-VM config file if one is found"
+	echo "    halt                 Alias to --destroy-only"
+	echo "    wipe                 Alias to --stop-only"
+	echo "    update               Alias to --no-delete --no-install-dev --no-restore-saved"
+	echo "                         and disables user/group account sync from host to container"
+	echo "    reboot               Alias to update with --no-download"
 	echo "    -h|--help            print this help"
 }
 
@@ -229,6 +237,7 @@ DOTDOMAINNAME=""
 [ -z "$ALLOW_CONFIG_FILE" ] && ALLOW_CONFIG_FILE=yes
 [ -z "${OVERLAYFS-}" ] && OVERLAYFS="auto"
 [ -z "${NO_RESTORE_SAVED-}" ] && NO_RESTORE_SAVED=no
+[ -z "${NO_DELETE-}" ] && NO_DELETE=no
 
 while [ $# -gt 0 ] ; do
 	case "$1" in
@@ -257,11 +266,11 @@ while [ $# -gt 0 ] ; do
 		[ -z "$2" ] && APT_PROXY="-" || APT_PROXY="$2"
 		shift 2
 		;;
-	--stop-only)
+	--stop-only|--wipe-only|wipe)
 		STOPONLY=yes
 		shift
 		;;
-	--destroy-only)
+	--destroy-only|--halt-only|halt)
 		DESTROYONLY=yes
 		shift
 		;;
@@ -283,6 +292,28 @@ while [ $# -gt 0 ] ; do
 		;;
 	--deploy-only)
 		DEPLOYONLY=yes
+		shift
+		;;
+	--no-delete)
+		NO_DELETE=yes
+		shift
+		;;
+	--with-delete)
+		NO_DELETE=no
+		shift
+		;;
+	reboot) # New uptime for existing rootfs - no initial reconfigs to do now
+		ATTEMPT_DOWNLOAD=no
+		;&
+	update) # New uptime for existing rootfs - no initial reconfigs to do now
+		NO_DELETE=yes
+		NO_RESTORE_SAVED=yes
+		INSTALL_DEV_PKGS=no
+		FORCE_RUN_APT=""
+		export FORCE_RUN_APT
+		COPYHOST_USERS=""
+		COPYHOST_GROUPS=""
+		ELEVATE_USERS=""
 		shift
 		;;
 	--attempt-download)
@@ -364,7 +395,7 @@ fi
 # Note: several hardcoded paths are expected relative to "snapshots", so
 # it is critical that we succeed changing into this directory in the end.
 
-mkdir -p "/srv/libvirt/rootfs" "/srv/libvirt/overlays"
+mkdir -p "/srv/libvirt/rootfs" "/srv/libvirt/overlays" "/srv/libvirt/overlays-ro"
 cd "/srv/libvirt/rootfs" || \
 	die "Can not 'cd /srv/libvirt/rootfs' to keep container root trees"
 
@@ -431,6 +462,12 @@ xno)
 		"into a dedicated full RW directory"
 	;;
 esac
+
+if [ "$NO_DELETE" = yes ] ; then
+	if [ x"$OVERLAYFS" = xno ] && [ x"$ATTEMPT_DOWNLOAD" != xno ] ; then
+		die "Requested to not delete VM contents and to download a new tarball root at the same time"
+	fi
+fi
 
 # Verify that this script runs once at a time for the given VM
 if [ -n "$VM" ] && [ -f "$VM.lock" ] ; then
@@ -624,7 +661,7 @@ else
 			ensure_md5sum "$IMAGE" "$IMAGE.md5" || IMAGE=""
 		done
 	fi
-	IMAGE_FLAT="`basename "$IMAGE" .$EXT`_${IMGTYPE}_${ARCH}.$EXT"
+	IMAGE_FLAT="`basename "$IMAGE" .$EXT`_${IMGTYPE}_${ARCH}_${IMGQALEVEL}.$EXT"
 fi
 if [ -z "$IMAGE" ]; then
 	die "No downloaded image files located in my cache (`pwd`/$IMGTYPE/$ARCH/*.$EXT)!"
@@ -650,7 +687,8 @@ sleep 5
 ALTROOT="$(cd "`pwd`/../rootfs/$VM" && pwd)" || die "Could not determine the container ALTROOT"
 logmsg_info "Unmounting paths related to VM '$VM':" \
 	"'${ALTROOT}/lib/modules', '${ALTROOT}/root/.ccache'" \
-	"'${ALTROOT}/proc', '${ALTROOT}', '`pwd`/../rootfs/${IMAGE_FLAT}-ro'"
+	"'${ALTROOT}/proc', '${ALTROOT}', '`pwd`/../rootfs/${IMAGE_FLAT}-ro'," \
+	"'`pwd`/../overlays-ro/${IMAGE_FLAT}-ro'"
 umount -fl "${ALTROOT}/lib/modules" 2> /dev/null > /dev/null || true
 umount -fl "${ALTROOT}/root/.ccache" 2> /dev/null > /dev/null || true
 umount -fl "${ALTROOT}/proc" 2> /dev/null > /dev/null || true
@@ -659,6 +697,7 @@ fusermount -u -z "${ALTROOT}" 2> /dev/null > /dev/null || true
 
 # This unmount can fail if for example several containers use the same RO image
 # or if it is not used at all; not shielding by "$OVERLAYFS" check just in case
+umount -fl "../overlays-ro/${IMAGE_FLAT}-ro" 2> /dev/null > /dev/null || true
 umount -fl "../rootfs/${IMAGE_FLAT}-ro" 2> /dev/null > /dev/null || true
 
 # root bash history may be protected by chattr to be append-only
@@ -670,28 +709,55 @@ if [ x"$DESTROYONLY" = xyes ]; then
 	exit 0
 fi
 
-# Destroy the overlay-rw half of the old running container, if any
-if [ -d "../overlays/${IMAGE_FLAT}__${VM}" ]; then
-	logmsg_info "Removing RW directory of the stopped VM:" \
-		"'../overlays/${IMAGE_FLAT}__${VM}'"
-	rm -rf "../overlays/${IMAGE_FLAT}__${VM}"
-	rm -rf "../overlays/${IMAGE_FLAT}__${VM}.tmp"
-	sleep 1; echo ""
-fi
+if [ x"$NO_DELETE" = xyes ] ; then
+	logmsg_info "NO_DELETE==yes, not deleting old VM overlay data"
+	# Just in case, do not check OVERLAYFS envvar here too
 
-# When the host gets ungracefully rebooted, useless old dirs may remain...
-for D in ../overlays/*__${VM}/ ../overlays/*__${VM}.tmp/ ; do
-	if [ -d "$D" ]; then
-		logmsg_warn "Obsolete RW directory for an old version" \
-			"of this VM was found, removing '`pwd`/$D':"
-		ls -lad "$D"
-		rm -rf "$D"
+	OLD_OVERLAY_RW="`ls -1dt ../overlays/*__${VM}/ | head -1 | sed 's,/*$,,'`" && \
+	[ -n "$OLD_OVERLAY_RW" ] && \
+	if [ x"$OLD_OVERLAY_RW" != x"../overlays/${IMAGE_FLAT}__${VM}" ] \
+	&& [ ! -d "../overlays/${IMAGE_FLAT}__${VM}" ] \
+	; then
+		logmsg_info "Preserving '$OLD_OVERLAY_RW' as '../overlays/${IMAGE_FLAT}__${VM}'"
+		mv "$OLD_OVERLAY_RW" "../overlays/${IMAGE_FLAT}__${VM}"
+	fi
+
+	# When the host gets ungracefully rebooted, useless old dirs may remain...
+	for D in ../overlays/*__${VM}.tmp/ ; do
+		if [ -d "$D" ]; then
+			logmsg_warn "Obsolete RW tmp directory for an old version" \
+				"of this VM was found, removing '`pwd`/$D':"
+			ls -lad "$D"
+			rm -rf "$D"
+			sleep 1; echo ""
+		fi
+	done
+else
+	# Destroy the overlay-rw half of the old running container, if any
+	if [ -d "../overlays/${IMAGE_FLAT}__${VM}" ]; then
+		logmsg_info "Removing RW directory of the stopped VM:" \
+			"'../overlays/${IMAGE_FLAT}__${VM}'"
+		rm -rf "../overlays/${IMAGE_FLAT}__${VM}"
+		rm -rf "../overlays/${IMAGE_FLAT}__${VM}.tmp"
 		sleep 1; echo ""
 	fi
-done
-for D in ../rootfs/*-ro/ ; do
+
+	# When the host gets ungracefully rebooted, useless old dirs may remain...
+	for D in ../overlays/*__${VM}/ ../overlays/*__${VM}.tmp/ ; do
+		if [ -d "$D" ]; then
+			logmsg_warn "Obsolete RW directory for an old version" \
+				"of this VM was found, removing '`pwd`/$D':"
+			ls -lad "$D"
+			rm -rf "$D"
+			sleep 1; echo ""
+		fi
+	done
+fi
+
+for D in ../overlays-ro/*-ro/ ../rootfs/*-ro/ ; do
 	# Do not remove the current IMAGE mountpoint if we reuse it again now
-	[ x"$D" = x"../rootfs/${IMAGE_FLAT}-ro/" ] && continue
+	### [ x"$D" = x"../rootfs/${IMAGE_FLAT}-ro/" ] && continue
+	[ x"$D" = x"../overlays-ro/${IMAGE_FLAT}-ro/" ] && continue
 	# Now, ignore non-directories and not-empty dirs (used mountpoints)
 	if [ -d "$D" ]; then
 		# This is a directory
@@ -730,15 +796,24 @@ for D in ../rootfs/*-ro/ ; do
 	fi
 done
 
-# clean up VM space
-logmsg_info "Removing VM rootfs from '${ALTROOT}'"
-if ! /bin/rm -rf "${ALTROOT}" ; then
-	logmsg_error "FAILED to remove '${ALTROOT}'"
-	logmsg_info "Checking if blocked by any processes?.."
-	fuser "${ALTROOT}" "${ALTROOT}"/*
-	fuser -c "${ALTROOT}" "${ALTROOT}"/*
-	fuser -m "${ALTROOT}" "${ALTROOT}"/*
-	die "Can not manipulate '${ALTROOT}' at this time"
+if [ x"$OVERLAYFS" != xyes ] \
+&& [ x"$NO_DELETE" = xyes ] \
+&& [ x"$ATTEMPT_DOWNLOAD" = xno ] \
+; then
+	# For overlayfs mode, we do make sure the new rootfs dir is empty
+	# For tarball mode, we keep old root if the tarball did not change
+	logmsg_info "NO_DELETE==yes, not deleting old VM tarball data"
+else
+	# clean up VM space
+	logmsg_info "Removing VM rootfs from '${ALTROOT}'"
+	if ! /bin/rm -rf "${ALTROOT}" ; then
+		logmsg_error "FAILED to remove '${ALTROOT}'"
+		logmsg_info "Checking if blocked by any processes?.."
+		fuser "${ALTROOT}" "${ALTROOT}"/*
+		fuser -c "${ALTROOT}" "${ALTROOT}"/*
+		fuser -m "${ALTROOT}" "${ALTROOT}"/*
+		die "Can not manipulate '${ALTROOT}' at this time"
+	fi
 fi
 
 if [ x"$STOPONLY" = xyes ]; then
@@ -750,9 +825,9 @@ fi
 logmsg_info "Creating a new VM rootfs at '${ALTROOT}'"
 mkdir -p "${ALTROOT}"
 if [ x"$OVERLAYFS" = xyes ]; then
-	logmsg_info "Mount the common RO squashfs at '`pwd`/../rootfs/${IMAGE_FLAT}-ro'"
-	mkdir -p "../rootfs/${IMAGE_FLAT}-ro"
-	mount -o loop "$IMAGE" "../rootfs/${IMAGE_FLAT}-ro" || \
+	logmsg_info "Mount the common RO squashfs at '`pwd`/../overlays-ro/${IMAGE_FLAT}-ro'"
+	mkdir -p "../overlays-ro/${IMAGE_FLAT}-ro"
+	mount -o loop "$IMAGE" "../overlays-ro/${IMAGE_FLAT}-ro" || \
 		die "Can't mount squashfs"
 
 	logmsg_info "Use the individual RW component" \
@@ -761,7 +836,7 @@ if [ x"$OVERLAYFS" = xyes ]; then
 	mkdir -p "../overlays/${IMAGE_FLAT}__${VM}" \
 		"../overlays/${IMAGE_FLAT}__${VM}.tmp"
 	mount -t ${OVERLAYFS_TYPE} \
-		-o lowerdir="../rootfs/${IMAGE_FLAT}-ro",upperdir="../overlays/${IMAGE_FLAT}__${VM}",workdir="../overlays/${IMAGE_FLAT}__${VM}.tmp" \
+		-o lowerdir="../overlays-ro/${IMAGE_FLAT}-ro",upperdir="../overlays/${IMAGE_FLAT}__${VM}",workdir="../overlays/${IMAGE_FLAT}__${VM}.tmp" \
 		${OVERLAYFS_TYPE} "${ALTROOT}" \
 	|| die "Can't overlay-mount rw directory"
 else
